puppetlabs-windows proof of concept
This commit creates a proof of concept for a windows collection
meta module. It declares dependencies on our supported/approved
Windows content, bound to the major version they're at now. The
only other thing it provides is a README which serves for the
module page on Forge. It's meant to describe the functionality
provided by the collection, with example usage and instructions on
how to get started. Much of it could and should supplement proper
solution-oriented material on puppetlabs.com