(maint) Replace ns-name with ns/name in dependencies
Users have reported that in current versions of the Puppet module tool
when dependencies are specified in the form namespace-modulename the
command `puppet module list` will complain of missing dependencies.
Replacing the dependencies with namespace/modulename in the
metadata.json resolves this issue.

The source bug is described in PUP-3121 and resolved in Puppet 4. Users
of older versions could still run into the undesirable behavior though
so it still makes sense to update the metadata.json file here for now.